+++ patrick_logger.py
@@ -10,7 +10,7 @@ feedback if anything occurs to you.
 
 http://patrickbrianmooney.nfshost.com/~patrick/
 
-v1, 7 October 2015. 
+v1.1, 6 December 2015. 
 """ 
 
 # Can set the starting level above zero explicitly when debugging, esp. when debugging command-line options
@@ -20,8 +20,7 @@ def log_it(message,minimum_level=1):
     """Add a message to the log if verbosity_level is at least minimum_level.
     Currently, the log goes to standard output.
     """
-    if verbosity_level >= 4: # set verbosity to at least 4 to get this message output in the debug log
+    if verbosity_level >= 6: # set verbosity to at least 6 to get this message output in the debug log
         print("\nDEBUGGING: function log_it() called")
     if verbosity_level >= minimum_level:
         print(message)
-