+++ HISTORY.md
@@ -8,10 +8,18 @@ v1: 7 October 2015
 ------------------
 Only contains the file "patrick_logger.py", a simple logging application.
 
+v1.1: 28 December 2015
+----------------------
+* Small updates to `patrick_logger.py` to make it compatible with Python 3.
+* Added files `README.md` and `LICENSE.md`. It's totally, like, a real Git thing now!
+* Added `genClassCalendar.py`, a small script to generate vEvent markup for my syllabus every quarter.
+* Added `postprocess-photos.py`, a script to automate processing for photos after I offload them from my camera.
+
 FUTURE PLANS
 ============
-* Abstractify the logger interface in patrick_logger.py
+* Abstractify the logger interface in `patrick_logger.py`.
+* Do pure Python EXIF-based image roatation in `postprocess-photos.py`.
 
 KNOWN BUGS
 ==========
-* None known at the moment.
+* No bugs known at the moment.